PANIC if the workfile list in shared-memory is inconsistent(#10075)
workfile_shared->num_active may not the actual size of the
list workfile_shared->activeList in production. The best way
is to find the root cause and fix the inconsistency. However,
we failed to find where the problem code is. The above two
variables are both in shared memory. It's commonly to reset
the shared memory when part of the shared memory is corrupted.

Reviewed-by: Hao Wang <haowang@pivotal.io>
(cherry picked from commit 4c7854ee5545a3475e54baa06ed3b694aa16ca46)