PXF: fix glob matching in PXF paths for Hadoop-compatible filesystems
When querying a PXF external table with some glob matching paths in the
LOCATION URI, Tomcat is rejecting the paths. In this commit, we remove
the `path` query string parameter as it has been deprecated and is no
longer used in the server, and it's causing issues with Tomcat because
the query string parameter is not URL-encoded.