Enable reognization on partition leaf node
* Enable reognization on partition leaf node 

We are not allowed to modify the distribution of partition leaf table
after gpdb6. Because the distibution of root table and leaf table must
be same. But if only reorg on the partition leaf table, it will not
cause this problem. So we enable reorg when the distibution was not be
changed.

* Ignore a test on gp_explain to fix the flaky case first

After merge the commit d525474047757865f88a2f0d6d8e43837e15f70d,
The test "explain SELECT * from information_schema.key_column_usage"
in gp_explain sometimes fails. I think the main reason for the failure
is that the catalog table involved in this query will change during the
test. So the query plan should change. Ignore this test first to fix
pipeline first.