Log fewer errors (#10100)
This is a continuation of commit 456b2b3 in GPORCA. Adding more errors to the list that
doesn't get logged in log file. We are also removing the code that writes to std::cerr,
generating a not very nice looking log message. Instead, add the info whether the error was
unexpected to another log message that we also generate.

The original commit on the master branch is fba77702.