Fix the logic in pg_lock_status() to keep track of which row to return.
The logic with 'whichrow' and 'whichresultset' introduced in commit
991273b2b3 was slightly wrong. The last row (or first? not sure) of each
result set was returned twice, and the corresponding number of rows at the
end of the last result set were omitted.

For example:

postgres=# select gp_segment_id, * from pg_locks;
 gp_segment_id |  locktype  | database | relation | page | tuple | virtualxid | transactionid | classid | objid | objsubid | virtualtransaction |  pid  |      mode       | granted | fastpath | mppsessionid | mppiswriter | gp_segment_id
---------------+------------+----------+----------+------+-------+------------+---------------+---------+-------+----------+--------------------+-------+-----------------+---------+----------+--------------+-------------+---------------
            -1 | relation   |    13200 |    11869 |      |       |            |               |         |       |          | 1/8                | 28748 | AccessShareLock | t       | t        |            6 | t           |            -1
            -1 | virtualxid |          |          |      |       | 1/8        |               |         |       |          | 1/8                | 28748 | ExclusiveLock   | t       | t        |            6 | t           |            -1
             0 | virtualxid |          |          |      |       | 1/7        |               |         |       |          | 1/7                | 28750 | ExclusiveLock   | t       | t        |            6 | t           |             0
             1 | virtualxid |          |          |      |       | 1/7        |               |         |       |          | 1/7                | 28751 | ExclusiveLock   | t       | t        |            6 | t           |             1
             1 | virtualxid |          |          |      |       | 1/7        |               |         |       |          | 1/7                | 28751 | ExclusiveLock   | t       | t        |            6 | t           |             1
(5 rows)

Note how the last row is duplicated. And the row for 'virtualxid' from
segment 2 is omitted.

I noticed this while working on the PostgreSQL v12 merge: the 'lock'
regression test was failing because of this. I'm not entriely sure why we
haven't seen failures on 'master'. I think it's pure chance that none of
the lines that the test prints have been omitted on 'master'. But because
that's been failing, I don't feel the need to add more tests for this.

(cherry picked from commit c2e5e4af781098f3e38487c1f4258f6e4a552813)