Fix ICW test if GPDB compiled without ORCA
We need to ignore the output when enabling/disabling an Orca xform, as
if the server is not compiled with Orca there will be a diff (and we
don't really care about this output).

Additionally, clean up unnecessaary/excessive setting of GUCs

Some of these gucs were on by default or only intended for a specific
test. Explicitly setting them caused them to appear at the end of
`explain verbose` plans, making the expected output more difficult to
match with if the server was built with/without Orca.