Distinguish between tuple deleted by split udpate and normal delete
Split update (distribution key update) in Greenplum is implemented by
deleting a tuple from original segment and inserting it into new
segment.  This is similar to an update to partition key column in
upstream PostgreSQL, which is handled by commit f16241bef7cc.  This
patch cherry-picks relevant parts of that commit.  The patch helps to
distinguish a tuple deleted as part of split update from a tuple
deleted by a regular delete operation.  Concurrent attempts to update or
delete a split-updated tuple need to be aware that the update chain for
this tuple cannot be traced because the updated tuple no longer resides
in the same heap table (it's moved to another segment).  Concurrent
updates to a split-updated tuple now result in an error.  Previously,
this case was treated like a delete and it would lead to wrong results.

Main difference between the upstream commit and this patch is the tuple
header change is not recorded in WAL.  Mirrors see this as regular heap
delete during replay and crash recovery would also replay it as regular
heap delete.  That is, the tuple header change is treated like a hint
bit change.  Tuple masking logic used by replica check and WAL
consistency checks is updated to mask the tuple header change.  We think
such a change is OK to make in Greenplum 6 because it only affects
concurrently running transactions.  Tuples affected by it will be seen
as deleted once the split-update transaction completes.  Therefore,
binary compatibility, including downgrade to older minor version, is not
affected.

Note that moving tuples from one partition to another is prohibited in
Greenplum 6.  That's why we can repurpose the trick from upstream for
split-update.

Fixes GitHub issue #8919.

Discussion: https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/kET4SJ1AHBA/gauiHO2WAwAJ

Co-authored-by: Zhenghua Lyu <kainwen@gmail.com>
Reviewed-by: Gang Xiong <gxiong@pivotal.io>