Force to update checkpoint timeline id in control data after promotion and before pg_rewind. (#10402)
After promotion there is a checkpoint but not an immediate one, so the
checkpoint timeline id in pg_control data might not be updated in time when
running pg_rewind which depends on that to decide the necessity of incremental
recovery.  Let's run checkpoint after promotion in tests so that later
pg_rewind will not bypass the incremental recovery.

This should be able to fix some long-existing annoying flaky pg_rewind tests.

Reviewed-by:  (Jerome)Junfeng Yang <jeyang@pivotal.io>
(cherry picked from commit e072dff7df700974905eb91a3781ded671d41bab)