Remove gphostcache
Gphostcache has numerous issues, and has been a pain point for some
time. For this reason, we are removing it.

This commit moves the useful function of gphostcache - the hostname
deduping - to the gparray class, where a list of deduped hostnames is
returned from gpArray.get_hostlist().

There is a FIXME of correctly adding hostname to a newly added or
recovered mirror. The hostname resolution from address was incorrect and
faulty in its logic - an IP address never requires a hostname associated
with it. However, the hostname field in gp_segment_configuration should
be populated somehow - we recommend a "hostname" field addition to any
configuration files that require it. For now, we simple set the
"hostname" to "address" which ultimately delivers the same functionality
as the gphostcache implementation.

Co-authored-by: Tyler Ramer <tramer@vmware.com>
Co-authored-by: Jamie McAtamney <jmcatamney@vmware.com>
(cherry picked from commit f61d35cdeacc2f97e234648e06788207fb9e53fc)