Fix cdbpath_dedup_fixup does not consider merge append path.
Greenplum use unique row id path as a candidate to implement semijoin.
It is introduced long before. But GPDB6 has upgraded the kernel
version to Postgres 9.4 and introduced many new path types and
new plan nodes, thus cdbpath_dedup_fixup failed to consider them.
Some typical issues are: https://github.com/greenplum-db/gpdb/issues/9427

On Master branch, Heikki's commit 9628a33 refactored this part of code
so that it is OK on master. And for 4X and 5X, we do not have many new
kinds of plannode and pathnode, it is also OK.

It is very hard to backport commit 9628a33 to 6X, there is no concept of
a Path's target list in 9.4. And to totally remove this kind of path
is too overkilling. So the policy is to fix them one bye one if reported.