ic-proxy: handle early coming BYE correctly
In a query that contains multiple init/sub plans, the packets of the
second subplan might be received while the first is still being
processed in the ic-proxy mode, this is because in ic-proxy mode a local
host handshake is used instead of the global one.

To distinguish the packets of different subplans, especially for the
early coming ones, we must stop handling on the BYE immediately, and
pass any unhandled early coming pkts to the successor or the
placeholder.

This fixes the random hanging during the ICW parallel group of
qp_functions_in_from.  No new test is added.

Co-authored-by: Hubert Zhang <hzhang@pivotal.io>
Co-authored-by: Ning Yu <nyu@pivotal.io>
(cherry picked from commit 79ff4e62a8cc34ed0a4996ad6d1b0dbaed3bee4b)