Check if RelOptInfo exits before replace subplan for grouping set query
In grouping set cases, we need to replace our private subplan into
plan tree. But for subplan's RelOptInfo will not create if the
subplan does not have any relaton in its fromlist.

Check pointer in simple_rte_array and simple_rel_array to promise
which is not null before replacement.