Export MASTER_DATA_DIRECTORY when calling gpconfig
MASTER_DATA_DIECTORY must be exported to before calling
gpconfig else it will fail to set the GUC

Allow setting DCA_VERSION_FILE to enable testing

Allow overriding the value of DCA_VERSION_FILE so that during testing it
can be manipulated. Also add a test for the same to ensure that DCA
configuration GUCs are set properly on the environment.