Fix postgres fdw on gpdb 6X (#10680)
* Fix postgres_fdw's libpq issue

When using posgres_fdw, it reports the following error:
unsupported frontend protocol 28675.0: server supports 2.0 to 3.0

root cause: Even if postgres_fdw.so is dynamic linked to libpq.so
which is compiled with the option -DFRONTEND, but when it's loaded
in gpdb and run, it will use the backend libpq which is compiled together
with postgres program and reports the error.

We statically link libpq into postgres_fdw and hide all the symbols
of libpq.a with --exclude-libs=libpq.a to make it uses the frontend
libpq.

As postgres_fdw is compiled as a backend without -DFRONTEND, and linked
to libpq which is a frontend, but _PQconninfoOption's length is
different between backend and frontend as there is a macro in it.
The backend's _PQconninfoOption has field connofs, but the frontend
doesn't. This leads to the crash of postgres_fdw. One solution is to
remove the FRONTEND macro in struct PQconninfoOption to make PQconninfoOption
is same in the backend and frontend, but that brings an ABI change. To avoid that,
we add the FRONTEND macro on including libpq header files, so that postgres_fdw
can process the libpq's variables returned by libpq.a's functions as frontend.

* Disable postgres_fdw on MaxOS