Fix crashes when a Values Scan needs to create "fake ctids".
When converting semi-join to inner-join, a distinct agg on ctid is added
above the hash-join node. But the fake ctids generated in Values Scan
were invalid, with offset number 0, which caused an assertion failure.

This patch is based on commit d8886cf9, which fixed the same issue for
Function Scans.

Co-authored-by: dh-cloud <60729713+dh-cloud@users.noreply.github.com>
Co-authored-by: Jesse Zhang <sbjesse@gmail.com>