ic-proxy: support hostname as proxy addresses
The GUC gp_interconnect_proxy_addresses is used to set the listener
addresses and ports of all the proxy bgworkers, only IP addresses were
supported previously, which is inconvenient to use.

Now we add the support for hostnames too, the IP addresses are also
supported.

Note that if a hostname is bound to a different IP at runtime, we must
reload the setting with the "gpstop -u" command.

Reviewed-by: Hubert Zhang <hzhang@pivotal.io>
(cherry picked from commit 2a1794bc47e58ea033e68ee1d8ef3ab2f999706f)