Replace docker-image with registry-image.
For context, the registry-image resource was introduced more than two
years ago as a drop-in replace for the docker-image resource (especially
when you're only fetching, not building). It's leaner, less
resource-intense, faster, and doesn't rely on spawning a Docker daemon.

Also swept up in this patch are four unused files that are left behind
by previous changes.

This is spiritually a cherry-pick of commit be6ff30fb2e2a74d
(greenplum-db/gpdb#10818).

See https://github.com/concourse/registry-image-resource for more.