Use planner for one test in "subselect"
Previously after the 9.2 merge, we had a hack that short-circuits the
execution of the join when ORCA is on. This hack stops working with the
Postgres 12 merge because the FIXME subquery gets executed on the QEs as
well rather than just on the QD. This patch simply turns ORCA off for
the query.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>
Co-authored-by: Alexandra Wang <lewang@pivotal.io>
(cherry picked from commit 891e77d124fec427ebdbb8e3b9c6e2b3992254d6)