Fix ORCA crash for certain group by/ distinct queries when using gp_use_legacy_hashops (#10810)
When gp_use_legacy_hashops is set, for plans that have an Aggregate,
ORCA might crash when computing the distribution spec. This occurs when
the grouping columns are of a type that have default hash distribution
opfamilies, but do not have legacy hash distribution opfamilies.

This happens because the distribution is done only for types for which
"IsRedistributable()" returns true. This is (incorrectly) determined by
looking (only) at the default opfamily (independent of the GUC).
Subsequently, CDistributionSpecHashed::PopulateDefaultOpfamilies() tries
to assemble opfamilies for all redistributable columns, which shouldn't
include such types (when the GUC is set), resulting in dereferencing a
nullptr.

For example, int[] type has a default distr opfamily, but no
legacy distr opfamily.

This commit fixes the implementation of IsRedistributable() to avoid the
above error.

Co-authored-by: Shreedhar Hardikar <hardikar@cs.wisc.edu>